fix: seg fault on search

+++ src/b-tree-buf.h
@@ -17,7 +17,7 @@ void populate_key(key *k,data_record *d, u16 rrn);
 
 void insert(b_tree_buf *b,io_buf *data,  data_record *d, u16 rrn);
 
-u16 insert_key(b_tree_buf *b, page *p, key k, key *promo_key, page *r_child);
+int insert_key(b_tree_buf *b, page *p, key k, key *promo_key, page *r_child);
 
 void split(page *p, key k, page *r_child, key *promo_key, page *new_page, int pos);
 
@@ -35,7 +35,7 @@ page *search(b_tree_buf *b, const char *s);
 
 u16 search_key(b_tree_buf *b, page *p, key key, u16 *found_pos, page *return_page);
 
-u16 search_in_page(page *page, key key, int *return_pos);
+int search_in_page(page *page, key key, int *return_pos);
 
 int remove_key( b_tree_buf*b, page *page);
